[C8] Split views.go per view kind (#2227)
model/drawing/views.go was 659 LOC, over the 500-line ceiling. Split the
drawing view generation per view kind, leaving the collection + shared
projection path in place; each file is now under the ceiling:

- views.go (236): the DrawingViews collection, the recompute (associativity)
  pass, the per-view projection-frame dispatch (basisFor) and label/style edits.
- views_base.go (133): base + projected views (specs, adders, previews, edits).
- views_section.go (69): section views.
- views_detail.go (61): detail views.
- views_derived.go (216): auxiliary, slice, breakout, break and draft views,
  plus the shared addDerived adder.

Behaviour-preserving: every struct and function body is byte-identical, only
moved; imports narrowed per file (fmt dropped where a file's error paths all
delegate to parentBaseFor/uniqueName). Drawing corpus + archguard pass before
and after with no golden updates. Net delta of recognizers / tolerance
constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2227.

(cherry picked from commit d3e51f797725e7e28073761679f493b0aedee26d)
